+++ rust/src/router/transport.rs
@@ -7,6 +7,7 @@ use crate::messages::{
     TransportEnableTraceEventRequest, TransportGetStatsRequest, TransportProduceDataRequest,
     TransportProduceRequest, TransportSetMaxIncomingBitrateRequest,
     TransportSetMaxOutgoingBitrateRequest,
+    TransportSetMinOutgoingBitrateRequest,
 };
 pub use crate::ortc::{
     ConsumerRtpParametersError, RtpCapabilitiesError, RtpParametersError, RtpParametersMappingError,
@@ -393,6 +394,12 @@ pub(super) trait TransportImpl: TransportGeneric {
     ) -> Result<(), RequestError> {
         self.channel()
             .request(self.id(), TransportEnableTraceEventRequest { types })
+            .await
+    }
+
+    async fn set_min_outgoing_bitrate_impl(&self, bitrate: u32) -> Result<(), RequestError> {
+        self.channel()
+            .request(self.id(), TransportSetMinOutgoingBitrateRequest { bitrate })
             .await
     }